requirements.txt
diffusers==0.25.0
streamlit==1.32.0
torch==2.1.0
transformers==4.36.2
python-dotenv==1.0.1
Pillow==10.2.0
